fix: update worldMatrix calculation in PBR shader
- Modified the worldMatrix calculation to use v_instanceIds.x directly, enhancing the clarity and correctness of the shader's logic related to instance handling.

+++ src/webgl/shaderity_shaders/PbrSingleShader/PbrSingleShader.frag
@@ -511,7 +511,7 @@ void main ()
 
     diffuseTransmissionProps.diffuseTransmissionThickness = 1.0;
   #ifdef RN_USE_VOLUME
-    mat4 worldMatrix = get_worldMatrix(uint(v_instanceIds.x));
+    mat4 worldMatrix = get_worldMatrix(v_instanceIds.x);
     diffuseTransmissionProps.diffuseTransmissionThickness = volumeProps.thickness * (length(worldMatrix[0].xyz) * length(worldMatrix[1].xyz) * length(worldMatrix[2].xyz)) / 3.0;
   #endif // RN_USE_VOLUME
